use new ActiveRecord::RuntimeRegistry for rails 7.1
This eliminates the need for monkey patching ActiveRecord::LogSubscriber.
